requirements.txt
schedule
vk_api
requests
python-dotenv
pytz